+++ android-data-core-sdk/src/main/java/io/fastpix/data/sdkBuild/SDKBuildConfig.kt
@@ -2,6 +2,6 @@ package io.fastpix.data.sdkBuild
 
 object SDKBuildConfig {
     const val SDK_NAME = "fastpix-data-core-sdk"
-    const val SDK_VERSION = "1.2.9"
+    const val SDK_VERSION = "1.3.0"
     var SDK_URL = ""
 }